Refactor PatientInfoTable
